% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstValide.R
\name{EstValide}
\alias{EstValide}
\title{Détermine les numéros possibles à un emplacement}
\usage{
EstValide(X, row, col)
}
\arguments{
\item{X}{la matrice}

\item{row}{la ligne}

\item{col}{la colonne}
}
\value{
la liste des nombres possibles
}
\description{
Détermine les numéros possibles à un emplacement
}
