% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible.R
\name{possible}
\alias{possible}
\title{Informe sur la validité du nombre dans la ligne/colonne/bloc}
\usage{
possible(X, row, col, number)
}
\arguments{
\item{X}{une matrice}

\item{row}{la ligne}

\item{col}{la colonne}

\item{number}{le nombre demandé}
}
\value{
True pour valide ou False pour non valide
}
\description{
Informe sur la validité du nombre dans la ligne/colonne/bloc
}
